Merge in add yourself to album
Couple of random shims here that need to be refactored later. But
basically, if viewing an album that osmeone else owns, you can add
yourself to that album as well by clicking the button. it will check to
see if the album already exists, and if so, add it to your default
collection.